First steps in segmentation
By considering the HSV color space we can clearly see the leaf,  in particular in the first two channels.

So by adding and suibtracting them up, we can find the contours through Otsu binarization.

Finally by subtracting the resulting contours the result should be only the leaf.

Binarizing the remaining contours we have a starting point for the leaf, since they're not fully connected.

NEED TESTING ON ALL LEAVES